+++ .idea/dictionaries/Pavel.xml
@@ -1,0 +1,7 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="Pavel">
+    <words>
+      <w>pamishenko</w>
+    </words>
+  </dictionary>
+</component>
