+++ src/main/java/ru/pamishenko/myBudget/Launcher.java
@@ -1,0 +1,7 @@
+package ru.pamishenko.myBudget;
+
+public class Loader {
+    public static void main(String[] args) {
+
+    }
+}
