+++ src/main/java/ru/pamishenko/myBudget/configs/AppConfig.java
@@ -1,0 +1,4 @@
+package ru.pamishenko.myBudget.configs;
+
+public class AppConfig {
+}
