+++ src/main/java/ru/pamishenko/myBudget/dao/UserDAO.java
@@ -1,0 +1,4 @@
+package ru.pamishenko.myBudget.dao;
+
+public interface UserDAO {
+}
