+++ src/main/java/ru/pamishenko/myBudget/model/goods/Cathegory.java
