+++ src/main/java/ru/pamishenko/myBudget/model/goods/Product.java
