+++ src/main/java/ru/pamishenko/myBudget/model/goods/Saler.java
