+++ src/main/java/ru/pamishenko/myBudget/model/goods/Type.java
