+++ src/main/java/ru/pamishenko/myBudget/model/person/Rolle.java
