+++ src/main/java/ru/pamishenko/myBudget/repositories/UserDAOImpl.java
@@ -1,0 +1,4 @@
+package ru.pamishenko.myBudget.repositories;
+
+public class UserDAOImpl {
+}
