+++ src/main/java/ru/pamishenko/myBudget/repositories/UsersRepository.java
@@ -1,0 +1,4 @@
+package ru.pamishenko.myBudget.repositories;
+
+public interface UsersRepository {
+}
