+++ src/main/java/ru/pamishenko/myBudget/repositories/myBudgetDAOImpl.java
@@ -1,0 +1,4 @@
+package ru.pamishenko.myBudget.repositories;
+
+public class myBudgetDAOImpl {
+}
