+++ src/main/java/ru/pamishenko/myBudget/services/RolleCRUD.java
@@ -6,6 +6,7 @@ import java.io.InputStreamReader;
 import org.hibernate.cfg.Configuration;
 import org.hibernate.SessionFactory;
 import org.hibernate.Session;
+import ru.pamishenko.myBudget.model.person.Rolle;
 
 
 public class RolleCRUD {
