+++ src/main/java/ru/pamishenko/myBudget/services/UserService.java
@@ -1,0 +1,4 @@
+package ru.pamishenko.myBudget.services;
+
+public interface UserService {
+}
