+++ src/main/java/ru/pamishenko/myBudget/services/UserServiceImpl.java
@@ -1,0 +1,4 @@
+package ru.pamishenko.myBudget.services;
+
+public class UserServiceImpl {
+}
