+++ src/main/webapp/WEB-INF/templates/lounge/about/about.html
@@ -1,0 +1,14 @@
+<html>
+    <head>
+        <meta http-equiv="Content-Type" content="text/html; charset=utf-8">
+        <title>about</title>
+    </head>
+    <body>
+        <h1>Добро пожаловать</h1>
+        <p>
+            Заходите каждый день, для <a href="index.html" content="ввести расходы">внесения расходов</a>.
+        </p>
+        <h2>27.09</h2>
+
+    </body>
+</html>
