+++ src/main/webapp/WEB-INF/templates/lounge/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Title</title>
+</head>
+<body>
+<p>sdrfgsdrfesrferwf</p>
+</body>
+</html>
